+++ lib/main.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: unused_import, duplicate_import
 
 import 'package:appxemphim/data/model/service.dart';
-import 'package:appxemphim/page/user/login.dart';
 import 'package:appxemphim/page/user/optionalaccount.dart';
 import 'package:appxemphim/page/paymentmethod/paymentmethodwidget.dart';
 import 'package:appxemphim/page/user/registers.dart';
