+++ lib/page/payment/paymentwidget.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: use_build_context_synchronously, unused_import
 
-import 'package:appxemphim/page/user/login.dart';
+
 import 'package:appxemphim/page/naviFrame.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
