+++ lib/page/setting.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: unused_import, camel_case_types, non_constant_identifier_names, sized_box_for_whitespace, avoid_unnecessary_containers, avoid_types_as_parameter_names
-
-import 'package:appxemphim/page/user/login.dart';
+import 'package:appxemphim/page/login.dart';
 import 'package:appxemphim/page/naviFrame.dart';
 import 'package:appxemphim/page/history/purchase/historyPurchase.dart';
 import 'package:flutter/material.dart';
